+++ src/main/java/cheongsan/domain/deposit/mapper/DepositAccountMapper.java
@@ -4,23 +4,22 @@ import cheongsan.domain.deposit.entity.DepositAccount;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Mapper
 public interface DepositAccountMapper {
+    // 예금계좌 정보 저장
     void insertDepositAccount(DepositAccount depositAccount);
 
-    DepositAccount findByUserIdAndAccountNumber(
-            @Param("userId") Long userId,
-            @Param("accountNumber") String accountNumber);
+    // 사용자ID와 계좌번호로 예금계좌 조회
+    DepositAccount findByUserIdAndAccountNumber(@Param("userId") Long userId, @Param("accountNumber") String accountNumber);
 
     List<DepositAccount> findByUserId(@Param("userId") Long userId);
 
-    void updateBalance(
-            @Param("id") Long id,
-            @Param("currentBalance") java.math.BigDecimal currentBalance);
+    // 예금계좌 잔액 업데이트
+    void updateBalance(@Param("accountId") Long accountId, @Param("balance") BigDecimal balance);
 
-    boolean isDepositAccountExists(
-            @Param("userId") Long userId,
-            @Param("accountNumber") String accountNumber);
+    // 예금계좌 존재 여부 확인
+    boolean isDepositAccountExists(@Param("userId") Long userId, @Param("accountNumber") String accountNumber);
 }
