+++ src/main/resources/mapper/DepositAccountMapper.xml
@@ -4,13 +4,18 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="cheongsan.domain.deposit.mapper.DepositAccountMapper">
 
-    <insert id="insertDepositAccount" parameterType="cheongsan.domain.deposit.entity.DepositAccount">
+    <insert id="insertDepositAccount" parameterType="cheongsan.domain.deposit.entity.DepositAccount"
+            useGeneratedKeys="true" keyProperty="id">
         INSERT INTO deposit_accounts (user_id, organization_code, account_number, current_balance)
         VALUES (#{userId}, #{organizationCode}, #{accountNumber}, #{currentBalance})
     </insert>
 
     <select id="findByUserIdAndAccountNumber" resultType="cheongsan.domain.deposit.entity.DepositAccount">
-        SELECT id, user_id, organization_code, account_number, current_balance
+        SELECT id,
+               user_id,
+               organization_code,
+               account_number,
+               current_balance
         FROM deposit_accounts
         WHERE user_id = #{userId}
           AND account_number = #{accountNumber}
@@ -24,14 +29,14 @@
 
     <update id="updateBalance">
         UPDATE deposit_accounts
-        SET current_balance = #{currentBalance}
-        WHERE id = #{id}
+        SET current_balance = #{balance}
+        WHERE id = #{accountId}
     </update>
 
     <select id="isDepositAccountExists" resultType="boolean">
-        SELECT EXISTS (SELECT 1
-                       FROM deposit_accounts
-                       WHERE user_id = #{userId}
-                         AND account_number = #{accountNumber})
+        SELECT COUNT(*) > 0
+        FROM deposit_accounts
+        WHERE user_id = #{userId}
+          AND account_number = #{accountNumber}
     </select>
 </mapper>
